+++ src/main/java/lotto/view/inputview/InputView.java
@@ -16,4 +16,10 @@ public class InputView {
         System.out.println("지난 주 당첨 번호를 입력해 주세요.");
         return scanner.nextLine();
     }
+
+    public static Integer inputBonusBall(){
+        scanner = new Scanner(System.in);
+        System.out.println("보너스 볼을 입력해 주세요.");
+        return scanner.nextInt();
+    }
 }
