+++ src/test/java/lotto/domain/WinningLottoTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.stream.Collectors;
@@ -8,15 +7,10 @@ import java.util.stream.IntStream;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-class WinningLottoTest {
-    WinningLotto actual;
-
-    @BeforeEach
-    void setUp() {
-        actual = WinningLotto.createWinningLotto(IntStream.rangeClosed(1, 6)
-                .mapToObj(number -> Integer.parseInt(String.valueOf(number)))
-                .collect(Collectors.toList()));
-    }
+public class WinningLottoTest {
+    public static WinningLotto actual = WinningLotto.createWinningLotto(IntStream.rangeClosed(1, 6)
+            .mapToObj(number -> Integer.parseInt(String.valueOf(number)))
+            .collect(Collectors.toList()));
 
     @Test
     void null_check() {
