+++ src/test/java/lotto/domain/lottomanager/LottoNumberTest.java
@@ -1,4 +1,4 @@
-package lotto.domain.lottofactory;
+package lotto.domain.lottomanager;
 
 import org.junit.jupiter.api.Test;
